requirements.txt
streamlit
streamlit-drawable-canvas
pillow
docxtpl
numpy
openpyxl
reportlab
docx2pdf
watchdog